Add Rewards and Resources contracts
Rewards contracts facilitates the distribuion of rewards to users. The
initial version awards flat rewards in a form of tokenized crafting
resources. Further versions should use randomness and adapt to the
various totems collaborating on the same quest.

Resources contracts is a simple ERC-1155 smart contract that provides
totem bearers with crafting resources.