Add start and join quest methods
Add the ability to start and join quests. The starting a quest creates a
quest instance that can be joined by others. Currently an unlimited
amount of Totems can join the instance (which poses a danger of a DoS
where there are too many Totems to receive the rewards in a block
limit).