Breakout attribute gas setting offsetting
As the attributes setting cost for the first time is non-neglibly higher
than subsequent settings, the gas cost offsetting is required, but the
access control for ERC-7508 still needs to be configured prior to the
offset. This is why the offsetting was removed from the constructor and
now has a dedicated function.